+++ Assets/Scripts/Interface/Bars/ManaBar.cs
@@ -37,6 +37,6 @@ public class ManaBar : Bar
     private void SetTextValue(float _current)
     {
         if(Value)
-            Value.SetText(_current + "/" + CharacterManager.Instance.Stats.Intelligence.MaxMana.Value.ToString());
+            Value.SetText((int)_current + "/" + ((int)CharacterManager.Instance.Stats.Intelligence.MaxMana.Value).ToString());
     }
 }
